% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.synpdx_fit}
\alias{print.synpdx_fit}
\title{Print a synpdx_fit summary}
\usage{
\method{print}{synpdx_fit}(x, ...)
}
\arguments{
\item{x}{synpdx_fit}

\item{...}{unused}
}
\description{
Print a synpdx_fit summary
}
