% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{package}
\name{synpdx-package}
\alias{synpdx}
\alias{synpdx-package}
\title{synpdx: NLME-based interaction analysis for PDX combinations}
\description{
Fits NLME tumor growth and drug-effect models with SAEM
}
\author{
\strong{Maintainer}: Junho Yoon \email{mothquitoman@gmail.com}

}
\keyword{internal}
