% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{synpdx_model_fit}
\alias{synpdx_model_fit}
\title{Internal: fit NLME growth–drug model(s) with SAEM and select by BIC}
\usage{
synpdx_model_fit(
  data,
  control,
  drug_a,
  drug_b,
  combo,
  alpha_eta = 0.01,
  seed = 990202
)
}
\description{
Internal: fit NLME growth–drug model(s) with SAEM and select by BIC
}
\keyword{internal}
