% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.synpdx_fit}
\alias{print.synpdx_fit}
\title{Print a synpdx_fit object}
\usage{
\method{print}{synpdx_fit}(x, ...)
}
\arguments{
\item{x}{A synpdx_fit object.}

\item{...}{Unused.}
}
\description{
Print a synpdx_fit object
}
