% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_boot.R
\name{synpdx_bootstrap}
\alias{synpdx_bootstrap}
\title{Bootstrap median S_norm (percentile and pivotal CIs)}
\usage{
synpdx_bootstrap(
  fit,
  data,
  control,
  drug_a,
  drug_b,
  combo,
  B = 100,
  seed = 202,
  cutoff = c("auto", "none"),
  out_dir = NULL
)
}
\description{
Bootstrap median S_norm (percentile and pivotal CIs)
}
