% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chisq.R
\name{synpdx_chisq_test}
\alias{synpdx_chisq_test}
\title{Group-level chi-square test with AD GOF on PIT}
\usage{
synpdx_chisq_test(Di_vec, combo_df, alpha = 0.05, out_dir = NULL)
}
\description{
Group-level chi-square test with AD GOF on PIT
}
