% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{synpdx_compute_Di}
\alias{synpdx_compute_Di}
\title{Compute per-subject distance D_i and combo predictions}
\usage{
synpdx_compute_Di(
  fit,
  data,
  control,
  drug_a,
  drug_b,
  combo,
  cutoff = c("auto", "none"),
  out_dir = NULL
)
}
\description{
Compute per-subject distance D_i and combo predictions
}
