% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{synpdx_model_fit}
\alias{synpdx_model_fit}
\title{Internal: Fit one or auto-select NLME model(s) with optional RE backward selection}
\usage{
synpdx_model_fit(
  dat,
  control,
  drug_a,
  drug_b,
  combo,
  seed = 990202,
  model = c("auto", "exp_const", "exp_decay", "gomp_const", "gomp_decay",
    "logistic_const", "logistic_decay"),
  select_random_effects = FALSE
)
}
\description{
Internal: Fit one or auto-select NLME model(s) with optional RE backward selection
}
\keyword{internal}
